refactor: Extract execute template and template functions to general utils
